+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/importer/helper/PortletConfigurationImporterHelper.java
@@ -14,7 +14,7 @@
 
 package com.liferay.layout.internal.importer.helper;
 
-import com.liferay.layout.page.template.importer.PortletPreferencesPortletConfigurationImporter;
+import com.liferay.layout.importer.PortletPreferencesPortletConfigurationImporter;
 import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.model.Portlet;
 import com.liferay.portal.kernel.portlet.PortletIdCodec;
